chore: split javascript folder to js and ts
