+++ src/generators/index.ts
@@ -3,3 +3,4 @@ export * from "./AbstractRenderer";
 
 export * from "./java";
 export * from "./javascript";
+export * from "./typescript";
