+++ src/generators/java/index.ts
@@ -1,1 +1,3 @@
 export * from "./JavaGenerator";
+export { JAVA_DEFAULT_PRESET } from "./JavaPreset";
+export type { JavaPreset } from "./JavaPreset";
