+++ src/generators/javascript/index.ts
@@ -1,2 +1,3 @@
-export * from "./javascript/JavaScriptGenerator";
-export * from "./typescript/TypeScriptGenerator";
+export * from "./JavaScriptGenerator";
+export { JS_DEFAULT_PRESET } from "./JavaScriptPreset";
+export type { JavaScriptPreset } from "./JavaScriptPreset";
