+++ src/generators/typescript/TypeScriptRenderer.ts
@@ -1,8 +1,8 @@
-import { AbstractRenderer } from "../../AbstractRenderer";
+import { AbstractRenderer } from "../AbstractRenderer";
 import { TypeScriptOptions } from "./TypeScriptGenerator";
 
-import { FormatHelpers } from "../../../helpers";
-import { CommonModel, CommonInputModel, Preset } from "../../../models";
+import { FormatHelpers } from "../../helpers";
+import { CommonModel, CommonInputModel, Preset } from "../../models";
 
 /**
  * Common renderer for TypeScript types
