+++ src/generators/typescript/index.ts
@@ -1,0 +1,3 @@
+export * from "./TypeScriptGenerator";
+export { TS_DEFAULT_PRESET } from "./TypeScriptPreset";
+export type { TypeScriptPreset } from "./TypeScriptPreset";
